test: 기존 ViewIntegrationTest에 있던 폴링 기능 테스트 이전
