+++ backend/src/test/java/harustudy/backend/integration/PollingIntegrationTest.java
@@ -1,7 +1,9 @@
 package harustudy.backend.integration;
 
 import harustudy.backend.participant.domain.Participant;
+import harustudy.backend.participant.domain.Step;
 import harustudy.backend.participant.dto.ParticipantResponse;
+import harustudy.backend.polling.dto.ProgressResponse;
 import harustudy.backend.study.domain.Study;
 import harustudy.backend.polling.dto.WaitingResponse;
 import harustudy.backend.testutils.EntityManagerUtil;
@@ -9,6 +11,9 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator.ReplaceUnderscores;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.Arguments;
+import org.junit.jupiter.params.provider.MethodSource;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.MediaType;
 import org.springframework.test.web.servlet.MvcResult;
@@ -17,8 +22,10 @@ import java.nio.charset.StandardCharsets;
 import java.util.List;
 import java.util.stream.Stream;
 
+import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.SoftAssertions.assertSoftly;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 @SuppressWarnings("NonAsciiCharacters")
@@ -76,5 +83,44 @@ class PollingIntegrationTest extends IntegrationTest {
             softly.assertThat(waitingResponse.participants().size()).isEqualTo(3);
             softly.assertThat(waitingResponse.participants()).containsExactlyInAnyOrderElementsOf(expected);
         });
+    }
+
+    @ParameterizedTest
+    @MethodSource("progressPollingParameters")
+    void 스터디_아이디로_진행_페이지_폴링을_진행할_수_있다(int progressCount, String expected) throws Exception {
+        //given
+        for (int i = 0; i < progressCount; i++) {
+            스터디_진행();
+        }
+
+        //when
+        MvcResult result = mockMvc.perform(
+                        get("/api/progress")
+                                .header(HttpHeaders.AUTHORIZATION, memberDto1.createAuthorizationHeader())
+                                .param("studyId", String.valueOf(study.getId())))
+                .andExpect(status().isOk())
+                .andReturn();
+
+        //then
+        String jsonResponse = result.getResponse().getContentAsString(StandardCharsets.UTF_8);
+        ProgressResponse response = objectMapper.readValue(jsonResponse,
+                ProgressResponse.class);
+
+        assertThat(response.progressStep()).isEqualTo(expected);
+    }
+
+    void 스터디_진행() throws Exception {
+        mockMvc.perform(
+                        post("/api/studies/{studyId}/next-step", study.getId())
+                                .header(HttpHeaders.AUTHORIZATION, memberDto1.createAuthorizationHeader()))
+                .andExpect(status().isNoContent());
+    }
+
+    private static Stream<Arguments> progressPollingParameters() {
+        return Stream.of(
+                Arguments.of(1, Step.PLANNING.name().toLowerCase()),
+                Arguments.of(2, Step.STUDYING.name().toLowerCase()),
+                Arguments.of(3, Step.RETROSPECT.name().toLowerCase())
+        );
     }
 }
